# Fix real matrix sqrt and log for edge cases