--- a/stdlib/LinearAlgebra/src/triangular.jl
+++ b/stdlib/LinearAlgebra/src/triangular.jl
@@ -2521,7 +2521,7 @@ Base.@propagate_inbounds function _sqrt_quasitriu_offdiag_block_2x2!(R, A, i, j)
     Rii = @view R[irange, irange]
     Rjj = @view R[jrange, jrange]
     Rij = @view R[irange, jrange]
-    if !(iszero(Rij) && all(isfinite, Rii) && all(isfinite, Rjj))
+    if !iszero(Rij) && !all(isnan, Rij)
         _sylvester_2x2!(Rii, Rjj, Rij)
     end
     return R