Fixes #40141, also ~eliminates an unnecessary scalar `sqrt`~ tries to reduce potential numerical issues in `_sqrt_real_2x2!`.

The added tests fail on master. Test 1 fails unless we make this change to `_sqrt_real_2x2!`. Test 2 also fails unless we make this change to `_sqrt_quasitriu_offdiag_block_2x2!`.

Edit: Test 3 tests the case where `NaN`s would cause `LAPACK.trsyl!` to error.