@oscardssmith 2021-03-23
How does this affect the performance?