@sethaxen 2021-03-23
> How does this affect the performance?

This version of `_sqrt_real_2x2!` takes about 5ns longer than on master (~35ns on master, ~40ns here). Using the same benchmark code and matrices as in https://github.com/JuliaLang/julia/pull/39973#issuecomment-796379587, `sqrt_quasitriu` takes about 16.5 μs on master and 17.2μs here.