@sethaxen 2021-03-23
With these changes, `_sqrt_real_2x2!` is ~40% slower. That being said, on a 20x20 matrix, `sqrt_quasitriu` is only ~4% slower. `sqrt` overflows and underflows much less with these changes, so I think this decrease in performance is worth it.

#### Updated Benchmarks

Setup:
```julia
using Random; Random.seed!(1)
F = schur(randn(20, 20)^2)
x = F.T[1:2,1:2] # 2x2 block
y = zero(x)
```

Before:
```julia
julia> @btime $(LinearAlgebra._sqrt_real_2x2!)($y, $x);
  35.130 ns (0 allocations: 0 bytes)
 
julia> @btime $(LinearAlgebra.sqrt_quasitriu)($(F.T));
  16.750 μs (1 allocation: 3.25 KiB)
```

This PR:
```julia
julia> @btime $(LinearAlgebra._sqrt_real_2x2!)($y, $x);
  50.037 ns (0 allocations: 0 bytes)

julia> @btime $(LinearAlgebra.sqrt_quasitriu)($(F.T));
  17.486 μs (1 allocation: 3.25 KiB)
```