@sethaxen 2021-03-24
The 2x2 diagonal block computation of `log` suffered from the same overflow/underflow, so I also adopted a similar approach for that.