@sethaxen 2021-03-24
With the recent changes, `_sqrt_real_2x2!` is almost as fast as the original:
```julia
julia> @btime $(LinearAlgebra._sqrt_real_2x2!)($y, $x);
  38.129 ns (0 allocations: 0 bytes)

julia> @btime $(LinearAlgebra.sqrt_quasitriu)($(F.T));
  17.354 μs (1 allocation: 3.25 KiB)
```