@sethaxen 2021-03-24
> This is a bugfix so this probably needs to be backported?

No, this fixes bugs introduced in #39973. The new tests added here, which fail on master, pass on Julia 1.6.